+++ src/de/tgx03/uno/host/Host.java
@@ -135,7 +135,7 @@ public class Host implements Runnable {
             }
         }
 
-        public void update(int[] cardCount) throws IOException {
+        public void update(short[] cardCount) throws IOException {
             Update update = new Update(game.getPlayer(this.id), game.getTopCard(), cardCount);
             output.writeObject(update);
         }
